% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regu-est.R
\name{ate.regu.path}
\alias{ate.regu.path}
\title{Model-assisted inference for average treatment effects along regularization paths}
\usage{
ate.regu.path(fold, nrho = NULL, rho.seq = NULL, y, tr, x,
  ploss = "cal", yloss = "gaus", off = NULL, ...)
}
\arguments{
\item{fold}{A vector of length 2, with the second component giving the fold number for cross validation in outcome regression. The first component is not used.}

\item{nrho}{A vector of length 2 giving the number of tuning parameters in a regularization path for PS estimation and that in cross validation for OR.}

\item{rho.seq}{A list of two vectors giving the tuning parameters for propensity score estimation (first vector) and outcome regression (second vector).}

\item{y}{An \eqn{n} x \eqn{1} vector of observed outcomes.}

\item{tr}{An \eqn{n} x \eqn{1} vector of treatment indicators (=1 if treated or 0 if untreated).}

\item{x}{An \eqn{n} x \eqn{p} matix of covariates, used in both propensity score and outcome regression models.}

\item{ploss}{A loss function used in propensity score estimation (either "ml" or "cal").}

\item{yloss}{A loss function used in outcome regression (either "gaus" for continuous outcomes or "ml" for binary outcomes).}

\item{off}{A \eqn{2} x \eqn{1} vector of offset values (e.g., the true values in simulations) used to calculate the z-statistics from augmented IPW estimation.}

\item{...}{Additional arguments to \code{\link{glm.regu.cv}} and \code{\link{glm.regu.path}}.}
}
\value{
\item{ps}{A list of 2 objects, giving the results from fitting the propensity score model by \code{\link{glm.regu.path}} for untreated (first) and treated (second).}
\item{mfp}{A list of 2 matrices of fitted propensity scores, along the PS regularization path, for untreated (first matrix) and treated (second matrix).}
\item{or}{A list of 2 lists of objects for untreated (first) and treated (second), where each object gives 
the results from fitting the outcome regression model by \code{\link{glm.regu.cv}} for a PS tuning parameter.}
\item{mfo}{A list of 2 matrices of fitted values from outcome regression based on cross validation, along the PS regularization path, 
for untreated (first matrix) and treated (second matrix).}
\item{est}{A list containing the results from augmented IPW estimation by \code{\link{ate.aipw}}.}
\item{rho}{A vector of tuning parameters leading to converged results in propensity score estimation.}
}
\description{
This function implements model-assisted inference for average treatment effects,
using regularized calibrated estimation along regularization paths for propensity score (PS) estimation 
while based on cross validation for outcome regression (OR).
}
\details{
See \strong{Details} for \code{\link{ate.regu.cv}}.
}
\examples{
\donttest{
data(simu.data)
n <- dim(simu.data)[1]
p <- dim(simu.data)[2]-2

y <- simu.data[,1]
tr <- simu.data[,2]
x <- simu.data[,2+1:p]
x <- scale(x)

ate.path.rcal <- ate.regu.path(fold=5*c(0,1), nrho=(1+10)*c(1,1), rho.seq=NULL, y, tr, x, 
                               ploss="cal", yloss="gaus")
ate.path.rcal$est
}

}
\references{
Tan, Z. (2017) Regularized calibrated estimation of propensity scores with model misspecification and high-dimensional data, arXiv:1710.08074. 

Tan, Z. (2019) Model-assisted inference for treatment effects using regularized calibrated estimation with high-dimensional data, 
\emph{Annals of Statistics}, to appear (preprint arXiv:1801.09817).
}
